# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "VBTeam_Record_Analysis_System API_DOC",
    "description": "API_DOC for VBTeam_Record_Analysis_System"
  },
  "host": "localhost:5000",
  "basePath": "",
...
  "schemes": [],
  "securityDefinitions": {},
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/update": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/home/calendar": {
      "get": {
        "tags": [
          "Home"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/home/calendar/create": {
      "post": {
        "tags": [
          "Home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/CreateEvent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "60d0b2011e44bec4e4be3a52"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/home/calendar/delete": {
      "post": {
        "tags": [
          "Home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "eventId": {
              "in": "body",
              "required": true,
              "type": "string"
            },
            "userId": {
              "in": "body",
              "required": true,
              "type": "string"
            },
            "in": "query",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "eventId": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/home/calendar/update": {
      "post": {
        "tags": [
          "Home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/UpdateEvent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/home/calendar/attend": {
      "post": {
        "tags": [
          "Home"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "eventId": {
                  "type": "string",
                  "example": "60d0b2011e44bec4e4be3a52"
                },
                "userId": {
                  "type": "string",
                  "example": "60d2cbeb9a8b913f2ef10193"
                },
                "attend": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/practice/posts": {
      "get": {
        "tags": [
          "Practice"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "type": "array",
            "schema": {
              "$ref": "#/definitions/Posts"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/practice/posts/upload": {
      "post": {
        "tags": [
          "Practice"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/UploadPost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/practice/posts/delete": {
      "post": {
        "tags": [
          "Practice"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "postId": {
              "in": "body",
              "required": true,
              "type": "string"
            },
            "userId": {
              "in": "body",
              "required": true,
              "type": "string"
            },
            "in": "query",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "postId": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/practice/posts/update": {
      "post": {
        "tags": [
          "Practice"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/UpdatePost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/match/records": {
      "get": {
        "tags": [
          "Match"
        ],
        "description": "",
        "parameters": [
          {
            "name": "recordId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "type": "array",
            "schema": {
              "$ref": "#/definitions/Record"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/match/records/create": {
      "post": {
        "tags": [
          "Match"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/CreateRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "60d619a88da34eda2a6ebc41"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/match/records/delete": {
      "post": {
        "tags": [
          "Match"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "recordId": {
                  "type": "string",
                  "example": "60d619a88da34eda2a6ebc41"
                },
                "userId": {
                  "type": "string",
                  "example": "60d08f760211c9a4925218a0"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Events": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60d0b2011e44bec4e4be3a52"
          },
          "start": {
            "type": "string",
            "example": "2021-06-21T18:00:00"
          },
          "end": {
            "type": "string",
            "example": "2021-06-21T20:00:00"
          },
          "title": {
            "type": "string",
            "example": "練球"
          },
          "place": {
            "type": "string",
            "example": "場一"
          },
          "attendance": {
            "type": "array",
            "example": [
              "60d2cbeb9a8b913f2ef10193"
            ],
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "example": "早點到！"
          }
        },
        "required": [
          "_id",
          "start",
          "title",
          "place"
        ]
      }
    },
    "CreateEvent": {
      "type": "object",
      "properties": {
        "event": {
          "type": "object",
          "properties": {
            "start": {
              "type": "string",
              "example": "2021-06-21T18:00:00"
            },
            "end": {
              "type": "string",
              "example": "2021-06-21T20:00:00"
            },
            "title": {
              "type": "string",
              "example": "練球"
            },
            "place": {
              "type": "string",
              "example": "場一"
            },
            "notes": {
              "type": "string",
              "example": "早點到！"
            }
          },
          "required": [
            "start",
            "title",
            "place"
          ]
        },
        "userId": {
          "type": "string",
          "example": "60d2cbeb9a8b913f2ef10193"
        }
      },
      "required": [
        "event",
        "userId"
      ]
    },
    "UpdateEvent": {
      "type": "object",
      "properties": {
        "event": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "60d0b2011e44bec4e4be3a52"
            },
            "start": {
              "type": "string",
              "example": "2021-06-21T18:00:00"
            },
            "end": {
              "type": "string",
              "example": "2021-06-21T20:00:00"
            },
            "title": {
              "type": "string",
              "example": "練球"
            },
            "place": {
              "type": "string",
              "example": "場一"
            },
            "notes": {
              "type": "string",
              "example": "早點到！"
            }
          },
          "required": [
            "_id",
            "start",
            "title",
            "place"
          ]
        },
        "userId": {
          "type": "string",
          "example": "60d2cbeb9a8b913f2ef10193"
        }
      },
      "required": [
        "event",
        "userId"
      ]
    },
...
    "Article": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "0d0b2011e44bec4e4be3a52"
        },
        "title": {
          "type": "string",
          "example": "6/24 練球筆記"
        },
        "content": {
          "type": "string",
          "example": "腳動起來！"
        },
        "uploader": {
          "type": "string",
          "example": "60d2cbeb9a8b913f2ef10193"
        },
        "uploadTime": {
          "type": "string",
          "example": "2021-06-23T22:15:55"
        }
      },
      "required": [
        "_id",
        "title",
        "content",
        "uploader",
        "uploadTime"
      ]
    },
    "UploadPost": {
      "type": "object",
      "properties": {
        "post": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "example": "6/24 練球筆記"
            },
            "content": {
              "type": "string",
              "example": "腳動起來！"
            },
            "uploadTime": {
              "type": "string",
              "example": "2021-06-23T22:15:55"
            }
          },
          "required": [
            "title",
            "content",
            "uploadTime"
          ]
        },
        "userId": {
          "type": "string",
          "example": "60d2cbeb9a8b913f2ef10193"
        }
      },
      "required": [
        "post",
        "userId"
      ]
    },
    "UpdatePost": {
      "type": "object",
      "properties": {
        "post": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "60d619a8840ea7a2a66d43a3"
            },
            "title": {
              "type": "string",
              "example": "~~~~~~~6/24 練球筆記"
            },
            "content": {
              "type": "string",
              "example": "~~~~~~~腳動起來！"
            },
            "uploadTime": {
              "type": "string",
              "example": "2021-06-23T22:15:55"
            }
          },
          "required": [
            "_id",
            "title",
            "content",
            "uploadTime"
          ]
        },
        "userId": {
          "type": "string",
          "example": "60d2cbeb9a8b913f2ef10193"
        }
      },
      "required": [
        "post",
        "userId"
      ]
    },
    "Posts": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Article"
      }
    },
    "Record": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "60d619a88da34eda2a6ebc41"
        },
        "type": {
          "type": "string",
          "example": "2020大資盃"
        },
        "team": {
          "type": "string",
          "example": "台大資工"
        },
        "opponent": {
          "type": "string",
          "example": "淡江資工"
        },
        "date": {
          "type": "string",
          "example": "2020-09-24"
        },
        "sets": {
          "type": "array",
          "example": [],
          "items": {}
        }
      },
      "required": [
        "_id",
        "type",
        "team",
        "opponent",
        "date"
      ]
    },
...
    "CreateRecord": {
      "type": "object",
      "properties": {
        "record": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "2020大資盃"
            },
            "team": {
              "type": "string",
              "example": "台大資工"
            },
            "opponent": {
              "type": "string",
              "example": "淡江資工"
            },
            "date": {
              "type": "string",
              "example": "2020-09-24"
            }
          },
          "required": [
            "type",
            "team",
            "opponent",
            "date"
          ]
        },
        "userId": {
          "type": "string",
          "example": "60d08f760211c9a4925218a0"
        }
      },
      "required": [
        "record",
        "userId"
      ]
    },
...
  }
}